ISAICP-2266-BACKPORT: Move services to core. Update references.
