+++ web/modules/custom/joinup_core/joinup_core.services.yml
@@ -2,3 +2,12 @@ services:
   joinup_core.workflow_access:
     class: Drupal\joinup_core\JoinupWorkflowAccessControlHandler
     arguments: ['@entity_type.manager', '@og.membership_manager']
+  joinup_core.relations_manager:
+    class: Drupal\joinup_core\JoinupRelationManager
+    arguments: ['@og.membership_manager']
+  joinup_core.workflow.user_provider:
+    class: Drupal\joinup_core\WorkflowUserProvider
+    arguments: ['@current_user']
+  joinup_core.fulfillment_guard:
+    class: Drupal\joinup_core\Guard\JoinupGuard
+    arguments: ['@entity_type.manager', '@joinup_core.workflow.user_provider', '@joinup_core.relations_manager', '@og.membership_manager', '@config.factory', '@current_user']
