+++ web/modules/custom/joinup_discussion/joinup_discussion.info.yml
@@ -5,6 +5,7 @@ core: 8.x
 package: Joinup
 dependencies:
   - joinup_core
+  - joinup_user
 
 config_devel:
   - core.base_field_override.node.discussion.status
